switch to react from jsxte
Fix issues with button activations and listeners

Use debug/error for logging from the library to make it easier to ignore useless comments and harder to ignore real issues.

Pre-emptively give errors for all types of triggers when they are specified but invalid.

Move triggers from option to the select they should be on - defaultValue is now used over selected b/c react :)